moved components in packages as named by vaadin
