+++ src/main/java/org/vaadin/firitin/button/DeleteButton.java
@@ -8,8 +8,8 @@ import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.html.Span;
 import org.claspina.confirmdialog.ButtonType;
 import org.claspina.confirmdialog.ConfirmDialog;
-import org.vaadin.firitin.components.VButton;
-import org.vaadin.firitin.components.VButton.ButtonSize;
+import org.vaadin.firitin.components.button.VButton;
+import org.vaadin.firitin.components.button.VButton.ButtonSize;
 import org.vaadin.firitin.fluency.ui.FluentComponent;
 import org.vaadin.firitin.fluency.ui.FluentHasEnabled;
