+++ src/main/java/org/vaadin/firitin/components/datepicker/VDatePicker.java
@@ -1,4 +1,4 @@
-package org.vaadin.firitin.components;
+package org.vaadin.firitin.components.datepicker;
 
 import com.vaadin.flow.component.AbstractField;
 import com.vaadin.flow.component.datepicker.DatePicker;
