+++ src/main/java/org/vaadin/firitin/components/html/DynamicFileDownloader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.firitin.components;
+package org.vaadin.firitin.components.html;
 
 import com.vaadin.flow.component.DetachEvent;
 import com.vaadin.flow.component.UI;
