+++ src/main/java/org/vaadin/firitin/components/radiobutton/VRadioButtonGroup.java
@@ -1,4 +1,4 @@
-package org.vaadin.firitin.components;
+package org.vaadin.firitin.components.radiobutton;
 
 import com.vaadin.flow.component.AbstractField.ComponentValueChangeEvent;
 import com.vaadin.flow.component.Component;
