+++ src/main/java/org/vaadin/firitin/components/upload/UploadFileHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.firitin.components;
+package org.vaadin.firitin.components.upload;
 
 import com.vaadin.flow.component.upload.Receiver;
 import com.vaadin.flow.component.upload.Upload;
