+++ src/test/java/org/vaadin/firitin/SimpleFieldsView.java
@@ -1,9 +1,8 @@
 package org.vaadin.firitin;
 
-import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.router.Route;
-import org.vaadin.firitin.components.VFormLayout;
+import org.vaadin.firitin.components.formlayout.VFormLayout;
 import org.vaadin.firitin.fields.CommaSeparatedStringField;
 
 import java.util.Arrays;
