+++ src/test/java/org/vaadin/firitin/UploadFileHandlerExample.java
@@ -22,7 +22,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import org.vaadin.firitin.components.UploadFileHandler;
+import org.vaadin.firitin.components.upload.UploadFileHandler;
 
 /**
  *
